Extract the scaleImage function to plotHelper folder, update all other file to import it;
